Combine the tests
기존 actions test와 reducer test를 하나의 파일로 합쳤습니다.
